#35 refactor: make PTabLayer contents unit to data
